requirements.txt
click==8.0.3
Flask==2.0.2
Flask-Cors==3.0.10
fuzzywuzzy==0.18.0
img2pdf==0.4.3
importlib-metadata==4.8.2
itsdangerous==2.0.1
Jinja2==3.0.3
joblib==1.0.1
lxml==4.6.4
MarkupSafe==2.0.1
numpy==1.21.4
pandas==1.2.4
pdf2image==1.16.0
pikepdf==4.0.1.post1
Pillow==8.4.0
psutil==5.8.0
pytesseract==0.3.8
python-dateutil==2.8.1
python-Levenshtein==0.12.2
pytz==2021.1
scikit-learn==0.24.2
scipy==1.7.2
scitime==0.0.2
six==1.16.0
threadpoolctl==2.1.0
typing-extensions==3.10.0.2
Unidecode==1.3.2
Werkzeug==2.0.2
wget==3.2
zipp==3.6.0
